Fix `update_graph_neighbors!()` to not update unsolvable graph aliases
`update_graph_neighbors!()` was previously invoked separately on `graph`
and `solvable_graph`, however we want to only update the solvable graph
aliases in `solvable_graph`, not the unsolvable ones.  We do this by
taking the setdiff of the aliases in `graph` and those in
`solvable_graph`, and only updating the solvable ones in
`solvable_graph`.